+++ tests/test_definitions/test_helpers.py
@@ -1,5 +1,3 @@
-from pandas import Timestamp
-
 from async_v20.definitions.base import Array
 from async_v20.definitions.helpers import flatten_dict
 from async_v20.definitions.primitives import AccountID, TradeID
@@ -13,20 +11,6 @@ flattened_dict = {'a_b': 2, 'a_c_d': 4}
 def test_flatten_dict():
     result = flatten_dict(nested_dict, delimiter='_')
     assert result == flattened_dict
-
-
-# def test_time_to_time_stamp_is_RFC3339_UNIX_agnostic():
-#     assert time_to_time_stamp('2017-11-19T23:47:30.000000003Z') == time_to_time_stamp('1511135250.000000003')
-#
-#
-# def test_unix_to_datetime64():
-#     assert type(time_to_time_stamp('1510937865.000000000')) == \
-#            Timestamp
-#
-#
-# def test_rfc3339_to_datetime64():
-#     assert type(time_to_time_stamp('2017-11-18T03:57:45.000000000')) == \
-#            Timestamp
 
 
 def test_get_valid_primitive_data_returns_primitive_example():
@@ -46,7 +30,7 @@ def test_get_valid_primitive_data_returns_default_value():
 
 
 def test_get_valid_primitive_data_returns_Array_example():
-    class TestArray(Array):
-        _contains = AccountID
+    class TestArray(Array, contains=AccountID):
+        pass
 
     assert (AccountID.example,) == get_valid_primitive_data(TestArray)